feat: Se implemento flujo para crear ventas. Se agregaron beans de configuracion. Se corrigio formato de data. Se agrego archivo docker-compose.yml para levantar bdd. Se agrego config de bdd en applitacion.yml.

+++ src/main/java/com/burgerreal/burger_gestion/domain/model/MetodoPago.java
@@ -7,7 +7,7 @@ public record MetodoPago (
     Long id,
     String nombre,
     BigDecimal porcentajeComision,
-    BigDecimal comisionFija,
+    Long comisionFija,
     boolean estaActivo
 ){
 }
